ml/requirements.txt
numpy
pandas
pandas_ta
scikit-learn
joblib
fastapi
uvicorn
flaskflask
